Refactor complet pour ajouter les DTO + résolution des requetes API

+++ src/main/java/com/fittrack/fit_track/service/PostService.java
@@ -1,12 +1,14 @@
 package com.fittrack.fit_track.service;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.fittrack.fit_track.dto.PostDTO;
+import com.fittrack.fit_track.mapper.PostMapper;
 import com.fittrack.fit_track.model.Post;
+import com.fittrack.fit_track.model.User;
 import com.fittrack.fit_track.repository.PostRepository;
 
 @Service
@@ -15,26 +17,41 @@ public class PostService {
     @Autowired
     private PostRepository postRepository;
 
-    public List<Post> getAllPosts() {
-        return postRepository.findAll();
+    public List<PostDTO> getAllPosts() {
+        List<Post> posts = postRepository.findAll();
+        return posts.stream()
+                    .map(PostMapper.INSTANCE::postToPostDTO)
+                    .toList();
+    }
+    
+    public PostDTO getPostById(Long id) {
+        Post post = postRepository.findById(id)
+            .orElseThrow(() -> new RuntimeException("Post not found"));
+        return PostMapper.INSTANCE.postToPostDTO(post);
     }
 
-    public Optional<Post> getPostById(Long id) {
-        return postRepository.findById(id);
+    public PostDTO createPost(PostDTO postDTO, User user) {
+        Post post = PostMapper.INSTANCE.postDTOToPost(postDTO);
+        post.setUser(user); // Associer l'utilisateur au post
+        Post savedPost = postRepository.save(post);
+        return PostMapper.INSTANCE.postToPostDTO(savedPost);
+    }
+       
+    // Mettre à jour un post à partir d'un DTO et retourner le DTO mis à jour
+    public PostDTO updatePost(Long id, PostDTO postDTO) {
+        Post post = postRepository.findById(id)
+            .orElseThrow(() -> new RuntimeException("Post not found"));
+        post.setContenu(postDTO.getContenu());
+        // Mettez à jour d'autres champs si nécessaire
+        Post updatedPost = postRepository.save(post);
+        return PostMapper.INSTANCE.postToPostDTO(updatedPost);
     }
 
-    public Post createPost(Post post) {
-        return postRepository.save(post);
-    }
-
-    public Post updatePost(Long id, Post postDetails) {
-        Post post = postRepository.findById(id).orElseThrow(() -> new RuntimeException("Post not found"));
-        post.setContenu(postDetails.getContenu());
-        // Modifier d'autres champs si nécessaire
-        return postRepository.save(post);
-    }
-
+    // Supprimer un post par ID
     public void deletePost(Long id) {
+        if (!postRepository.existsById(id)) {
+            throw new RuntimeException("Post not found");
+        }
         postRepository.deleteById(id);
     }
 }
